requirements.txt
streamlit
torch
opencv-python
pandas
numpy
Pillow
